Adjusts external apps to account for long app names
Converts external_tools.sass -> scss while adding a few styles to
truncate the names properly.  The styles added convert the table
to render using the fixed style and the first column to have a width
of 500px and overflow the text using ellipsis.  A title attribute is
also added to the column to help with UX.

fixes CNVS-7223

test plan:
	- Go to Settings > Apps
	- Add an app with a long name around 160 characters
	- You should see that the name is truncated with an ellipsis
	- Hover over the name
	- A tooltip with the full name should appear

Change-Id: I24cc71a0073799e66efa57421b80b0b70bdbecdb