update sis batch progress on an interval, default 2s
Rather than based on a # of rows processed metric

fixes CNVS-14640

test plan: run a large-ish SIS batch, and it should succeed as you'd
expect. the progress in the UI/API should still update as it runs (and
when it completes), but maybe not quite as often.

Change-Id: I44e30c1bbd1cd70ade7633c48389bf6bf0d04773
Reviewed-on: https://gerrit.instructure.com/39281
Reviewed-by: Zach Wily <zach@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>
QA-Review: Brian Palmer <brianp@instructure.com>
Tested-by: Brian Palmer <brianp@instructure.com>