Give good error message when user doesn't exist
fixes #CNVS-15119

This is for the workflow when you're adding
a user to a course as a teacher.  Currently
it makes it sound like the error is an email
formatting problem.

This also sets up a reusable way to push
account settings info over in the environment
and check it in a handlebars template.

TEST PLAN:
 - Login as an instructor
 - Make sure the root account has 
    "open_registration" turned off
 - Try to add a student to your course by email
     who isn't currently a canvas account holder
 - The error message should tell you that if the
     formatting is ok, you need to ask your admin
     if they have a canvas account


Change-Id: I0cfc07c4612ff4db7ef759cc146e612f94c1075b
Reviewed-on: https://gerrit.instructure.com/40230
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>