Quizzes only autosubmit when they hit the lock date
When a quiz is in progress, it should be marked late if the student is
still taking the quiz when the due date is reached.  If the lock date
comes up, the quiz should autosubmit at that point.  If there is a quiz
timer and a lock date, the lock date overrules the quiz timer.

Closes CNVS-13689

Test Plan:
  - Create 4 types of quiz scenarios:
    - Lock date only
      - submits at lock date
      - never marked late
    - Lock date and due date
      - autosubmits at lock date
      - late if submitted after due date
    - No due date
      - Never late
      - Autosubmits at course completion
    - Lock date and timer which runs beyond lock date
      - Gives the student time until lock, not normal timer
      - autosubmits at lock date
  - Monitor all these for sane prompts and alerts
    - deadline approaching flash messages for LATENESS (occur at T-30
      minutes, T-5 minutes, and T-1 minute)
    - autosubmission flash prompts (and popup @ T-10 seconds)

Change-Id: I5af9eb360a5f13527b32e7f84c205149bea1935b
Reviewed-on: https://gerrit.instructure.com/40609
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Trevor deHaan <tdehaan@instructure.com>
Product-Review: Ryan Taylor <rtaylor@instructure.com>