s/Login/Log In on mobile login page
fixes: CNVS-15370

When a user is logging in to Canvas, 
the Login button used to say Login. However, on 
the next screen when Canvas is requesting access 
to their account, the button says Log In. 

The second button is accurate in the context of 
how we are using the term Log In.

Change-Id: I3ad2d6cc86822f438e39b0d6e6089f2557224ec4
Reviewed-on: https://gerrit.instructure.com/40794
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>