add new ext_roles lti param for all the roles user has
this is a list of all the roles the user has in canvas
not just the ones for the current context.

This is how it should have been all along, but we can't
make the normal roles parameter do it this way because
tools depend on our current format. So send a new one
and tell people to  migrate over as they can

Test Plan:
 * do an lti launch, ext_roles should always have "user"
   and any of learner, instructure, administrator as appropriate

refs RD-518
closes PLAT-641

Change-Id: I8cf7e96ac88e1b17528cd3d717834762cc9b575b
Reviewed-on: https://gerrit.instructure.com/40875
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brad Humphrey <brad@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>