CQS: accessible score percentile chart
Closes CNVS-15301

Three enhancements:

  - chart now has a title (and apparently a browser tooltip!)
  - the 0% ... 100% X-Axis ticks are now hidden to the screenreader
  - the chart is equipped with an audible description, something like:
    "10 of the students scored above, or at, the average, and 2 below."

TEST PLAN
---- ----

  - go to CQS page
  - verify you hear the chart's title, "Score percentiles chart",
    labeled as a "group"
  - verify you can enter the group
    + verify that once you enter the group, you can hear the audible
      description of the chart noted above
  - verify you can not reach the 0%...100% X Axis ticks using the
    screen-reader

NVDA NOTES
---- -----

To get to the summary sentence that is now a replacement of the chart,
you can keep pushing DOWNARROW until you pass by all the table cells,
and just before it reads the "Question Breakdown" heading, or
alternatively (faster):

  - press H two times until you get to "Question Breakdown" then UPARROW
  - press T to get to the summary table, press DOWNARROW like 12 times
    to pass by all the cells, then another time

Something else: if the app renders before the data is loaded, e.g, cells
in the first table are still at 0%, NVDA might have already cached the
page. Then, when you get to the table and the app has loaded, it reads
"0%" instead of the actual numbers... you can use INSERT+F5 to tell NVDA
to refresh the page and look for new changes.

This is obviously problematic, but solving it is outside the scope of
this patch. Maybe we should prevent the app from starting until we have
all the data ready? Donno

See "Refresh browse mode document" in this section of their User Guide:
http://community.nvda-project.org/documentation/userGuide.html?#toc40

Change-Id: Ia4912bfb5978bf9c2fc2ae26d9a2967a5392e2f3
Reviewed-on: https://gerrit.instructure.com/41329
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Trevor deHaan <tdehaan@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>