fixes more of the a11y bugs @acannon found in newfiles
closes:  CNVS-15699

things fixed:

When adding a folder, the folder name input box is
not labeled. There is a placeholder, but no label
for screen reader users.

The column headers in the table I fear are too
verbose. This is a problem because they are read
each time you switch columns. I suggest removing
the unnecessary text that says "click to sort..."
to help shorten them up.

The toolbar, breadcrumbs, folder tree all have
regions, but the list of files and folder does not.
I suggest adding and naming a region for that
section as well.

After closing (canceling) the move modal, the focus
is not moved back to the item that opened it, or at
least to the general vacinity. This should be done,
whether a move was completed or not. Just being
dropped back at the beginning of the page is a
problem for keyboard only users.

The messaging that tells the user how many items
are selected is not automatically read, as I
believe it is supposed to.

All but the last three buttons on the toolbar are
unlabeled.

Change-Id: I0b856296ecc70f0d9e01a3d6859fe1079dfdb2e4
Reviewed-on: https://gerrit.instructure.com/41470
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Dan Minkevitch <dminkevitch@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>