SubmissionsController - download files when we're supposed to
When the link on the page says download, the user expects the result
to be that the file is downloaded rather than displayed inline.

The FilesController is responsible for the actual downloading,
and it uses a parameter called :download_frd that is set to
true to indicate a forced download instead of displaying inline.
When :download_frd is true, a special header is sent to the
browser instructing it to download the page.

Because the SubmissionsController receives the request for the file,
a redirect is made from the SubmissionsController to the
FilesController.  Previously the Subissionscontroller was omitting
the :download_frd parameter from the redirect.  This fixes that.

Test Plan:

1. Create an assignment that allows text type file submissions
2. As a student - submit an affected filetype (.txt, .c, .py, .html,
.css, etc)
3. As the instructor - view the submission in speedgrader
4. Click on "download the file" button and see that it now downloads
correctly

Fixes CNVS-8898

Change-Id: I568d8a52eb430895316cf04414e73e1081643e70
Reviewed-on: https://gerrit.instructure.com/41629
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>
Reviewed-by: Matt Berns <mberns@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>