make quiz question drag and drop to be useable by keyboard only users
This changes the quiz drag and drop handle to be the same icon we use for
drag and drop reordering on the index pages. Now when the icon is dragged it
gives normal drag and drop reordering, but when it is clicked it will bring
up the move modal dialog.

This also adds a warning message when the item is focused on that tells that it
is an accessible alternative.

This also fixes a bug with accessible reordering of quiz questions when there
are greater than 25 questions in a quiz.

fixes CNVS-15778

test plan:
  - as a teacher
  - create a quiz
    - the drag handle should now always show when creating questions and groups
    - you can drag the handle to reorder with drag-n-drop
    - you can click the handle to bring up the move to dialog box

  - create a quiz with no question groups
  - use screenreader
    - you should be able to use the dialog to reorder the questions
    - it should read the warning message on focus

  - create a quiz with > 25 questions
  - use screenreader
    - you should be able to use the dialog to reorder the questions
    - it should read the warning message on focus

  - create a quiz with question groups
  - use the screenreader
    - you should be able to reorder both questions/groups
    - try moving items in any way you can think - outside/inside groups, etc
    - it should read the warning message on focus

  - for all these situations, the normal drag-n-drop should continue to work
  - no drag and drop handles should show up in question banks
    - check banks with both < 50 questions and > 50 questions

Change-Id: If23ce2a68f1f8d60cc8f3ca76d4fea36f89b7a9b
Reviewed-on: https://gerrit.instructure.com/41703
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ryan Taylor <rtaylor@instructure.com>
QA-Review: Trevor deHaan <tdehaan@instructure.com>
Reviewed-by: Brian Finney <bfinney@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>