fix more a11y issues @acannon found
fixes: CNVS-15788

things this fixes:

The buttons in the toolbar are better, but
"restrict access" is still unlabeled.
Also "Downlod as Zip" is misspelled.

The tree no longer appears in the move dialog.

The list of files and folders still does not
appear to have a named landmark.

 "# items selected" is not picked up and read by screenreader

 things this doesn't try to fix:

 the thing about the checkboxes not actually becoming
 checked when clicked, that will be fixed in: CNVS-15785

 CNVS-15929 - make sure cog in newfiles has enough contrast

Change-Id: Ie8e5e0e4c93dfcd229ed73c940b24f7ab342e23b
Reviewed-on: https://gerrit.instructure.com/41706
Reviewed-by: Dan Minkevitch <dan@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>