Updated react-router
react-router wasn't correctly firing the
route handler in IE10, resulting in a blank
page.  This fixes that by updating react-router
to the newest version.  Note that some of the
functionality is currently broken in IE,
such as styles and the FolderTree.  Let's
do a run through of IE now that it's rendering
and make tickets of whatever is left to be fixed.

fixes CNVS-15930

test plan:
* Enable "Better File Browsing" feature flag
* Navigate to any course's Files tab in IE10
* The FilesApp should render and be semi-functional

Change-Id: I52e4ca73ace6ca0dec6b26a76c1a3cfa51e5d128
Reviewed-on: https://gerrit.instructure.com/42453
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Dan Minkevitch <dan@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>