make external tool list accessible
fixes CNVS-14932

Items should have been displayed in a link and list. They are now
displayed that way so should work in screen readers.

Test Plan

Given you have an external tool configured for a course
When you go to the modules page
And you add a new module item
When you select the "External Tool" dropdown option
Then you should see the external tool listed
And you should be able to click on it whith a mouse
And you should be able to select it with the screenreader and keyboard

Change-Id: I3ace34739c0eda3e45dc53df2d13f48cdf69077b
Reviewed-on: https://gerrit.instructure.com/44229
Reviewed-by: Ryan Shaw <ryan@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>