Color contrast - User avatars
fixes CNVS-16010

This changes the icon on the modal so that it meets accessibility
standards.

Test Plan:

- As a student go to your profile and click to change your icon
- The 'x' that appears on the modal now passes

- Go to styleguide and click on Modal Dialog > Show Dialog to see
the change in the global modal as well

Note for UI:
These modals are legacy and although still used throughout canvas
it felt best to replace the image with an svg than a png instead
of figuring out how to put in an <i> tag within the modal html.

Change-Id: Ia893ecc7d2f891dc92e8902060504292f67e10f9
Reviewed-on: https://gerrit.instructure.com/45353
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Chris Hart <chart@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Chris Hart <chart@instructure.com>