Enter now closes and opens agenda view items
fixes CNVS-16980

Both space and enter now close and open the agenda view item popup.

This change preserves the effect of Braden's original commit but
provides a way for the listener of the 'keyclick' event to stop
propagation and prevent default of the original underlying event that
caused the 'keyclick' event to be triggered by doing so on the
'keyclick' event object. Because of this, the scope of the test plan can
be much smaller than in previous patch sets.

Additionally, the behavior can now be added to
individual elements that don't have a role of button. This will prevent
accessibility errors where the role of button is applied to a header
which should have a role of header instead so that page navigation via a
keyboard isn't broken.

Test Plan:
* Go to the agenda view in the calendar and use your keyboard to open
  and close an event with both the enter key and the space key
* Ensure that you can toggle individual calendars and individual undated
  items with both enter and space key
* Go to the My Courses page. Using the keyboard navigate to a star and
  use both space bar and enter to toggle the star
* Go to Conversations and use the keyboard to get to the button next to
  the search field. Click it with both enter and space and make sure
  that it shows the dropdown menu

Change-Id: I432cba48231c3fdd64d9728e957ca697852ac8a2
Reviewed-on: https://gerrit.instructure.com/45379
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Braden Anderson <braden@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Matt Fairbourn <mfairbourn@instructure.com>