remove FileAPI swf files
refs CAT-680

These files are no longer in use. There is some more
code cleanup that can be done, but it will need to be
done in a couple stages, as there is a plugin that
depends on the trashable code which will need to be
updated first.

Change-Id: I3821d241664fbf74bbc5928e8615a71ad0a2ef47
Reviewed-on: https://gerrit.instructure.com/45625
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Paul Hinze <paulh@instructure.com>
QA-Review: Paul Hinze <paulh@instructure.com>
Product-Review: Paul Hinze <paulh@instructure.com>