fix formatting issue in firefox
fixes CNVS-17278

When you preview a file in firefox, the "info" button's icon is over
the info icon vs next to it. This fixes this by changing the button
into a link. This most likely should be a "button" I don' think usig a
link in this case is horrible so I went with it.

Test Plan
Given you are using firefox
And new files is enabled
And you have a file uplaoded
And you are currently previewing a file
Then you should see the "info" text is aligned to is icon
And It should still be aligned on all other browsers

Change-Id: I30cb0fbfca05ab0f1cb999211f8b9c6b494bbdc3
Reviewed-on: https://gerrit.instructure.com/45741
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Sterling Cobb <sterling@instructure.com>
QA-Review: Sterling Cobb <sterling@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>