use math to figure out if buttons are dark
this is just a little change so that you
don't have to tell the mixin if it's dealing with
a $is-light-background-button. it will figure it out
on it's own.  this is to make it more resilient
to accounts using their own branding

Change-Id: Ib6351dcc7428e5aadc739cee14f1994c8ff46b4f
Reviewed-on: https://gerrit.instructure.com/46161
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Chris Hart <chart@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>