canvas_update script: add double quotes and $() where apropo
bash best practice is to use double quotes to prevent globbing and word
splitting.  This probably doesn't matter in the context we are using,
but it pollutes the output of the linter so might as well just fix it.

Also switch `` to $().  Backticks are deprecated

Fixes CNVS-18183

Change-Id: I03cef6a7b3e1740bfb46c422e89788cd5d1eb7e0
Reviewed-on: https://gerrit.instructure.com/47726
Reviewed-by: Mark Severson <markse@instructure.com>
Tested-by: Jenkins
QA-Review: Adam Stone <astone@instructure.com>
Product-Review: Benjamin Porter <bporter@instructure.com>