lay groundwork for transitioning to smarter attachment verifiers
This commit adds a new form of attachment verifier. Rather than just
being the `uuid` from the attachment object, it is a signed JWT that
includes the attachment id, the viewing user id, and a viewing context.

The existing code that checks the verifier now checks to see if it's
a UUID and returns success if it is (to preserve legacy behavior). It
attempts to decode and verify the token, and verify that the user
captured in the token still has permission to :read or :download the
attachment.

The context captured in the token is currently unused, but may prove
useful for debugging or future purposes. The context is just an
asset_string for a model.

The only place generating new verifiers with this commit is the helper
ApplicationController#verified_file_download_url, which is only used by
content exports and eportfolios.

We also shoot increments to statsd when legacy and new verifiers are
used, so we can see when legacy verifier usage is low enough to disable.

Test plan:
 * Create a course export.
 * View the export at /courses/X/content_exports
 * Verify that the verifier= in the download URL is decoded correctly
   at jwt.io. It should have a user_id, an id, and a context.
 * Verify that a you can successfully download the export.
 * Use the files API and verify that urls still have the
   old-style verifier=<40-character UUID> on them, and that these
   links still work without authentication (such as an incognito tab
   or curl)

 * enable eportfolios in account settings
 * upload a file to user files
 * create an eportfolio, and attach a user file to it
 * save the eportfolio
 * confirm that the download link
   (which looks like /eportfolios/X/entries/Y/files/ZZZ)
   redirects to a file download link with a new verifier

fixes CNVS-18858
fixes CNVS-18861

Change-Id: I3db21469462dfdb8f8067f509a01a55d7cb23abd
Reviewed-on: https://gerrit.instructure.com/48637
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Jenkins
Reviewed-by: James Williams  <jamesw@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>