allow info button on file preview to get focus
the info button was changed to a link to fix a styling issue
on firefox (CNVS-17278), but since it didn't have an href, it
couldn't receive focus. this reverts it back to a button for
general accessibility goodness and fixes the styling issue.

also fixed an unintentional adding of "undefined" class to the
info button's classes when it was inactive.

fixes CNVS-18669

test plan:
 - preview a file in better file browsing
 - make sure the info button can get focus via tabbing
 - make sure the info button can be activated by keyboard
 - check the info button's styling in firefox

Change-Id: I439e3e1c9ff37f6af14bdb74898d5b90e3285b58
Reviewed-on: https://gerrit.instructure.com/49131
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
Product-Review: Dan Minkevitch <dan@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>