OutcomeView update calculation int intelligently
When the user changes between calculation methods, the calculation int
currently just stays the same.  This is almost never what we want.

Instead, this patch makes the calculation int update based on what is
selected.  When set to n_mastery, the value will change to 5 unless the
current value is already a valid settings for n_mastery.  When settings
to decaying_average, the value will change to 65 if it is currently set
to 5 (which is a valid but very unlikely value for decaying_average)

Fixes CNVS-18769

Test Plan:

    1. Edit an outcome on the outcomes page
    2. Set the value to decaying average and 65
    3. Change to n mastery and observe that the int changes to 5
    4. Change back to decaying average and observe that the int changes
    back to 65
    5. Change to n mastery and set 4
    6, Change to decaying average and observe the value stays at 4

Change-Id: Ie231bd2f108dfd896553e4a282422711bcb2003a
Reviewed-on: https://gerrit.instructure.com/49563
Reviewed-by: Matt Berns <mberns@instructure.com>
Tested-by: Jenkins
Product-Review: Hilary Scharton <hilary@instructure.com>
QA-Review: Adam Stone <astone@instructure.com>